Split ncurses wrapper into separate header file
This commit also removes the un-definition of the `assert` macro;
the ``assert.h`` header is included by the libncurses header, and in
order to define a function named ``assert`` in namespace ``nc`` there
was an ``#undef assert`` directive after including ``curses.h``. That
isn't really a good solution, so I've renamed the function to
``ncassert`` and removed the undef.

The current error-checking system is likely to change anyway, and the
new name is more search-replace friendly, which is good.